+++ lib/src/booleans.dart
@@ -1,5 +1,8 @@
 // TODO move into a util class https://github.com/jesims/docs/blob/master/Coding-Conventions/Dart.md#utilities
-/// Returns `true` if the boolean [value] is `false`
+import 'bool_extensions.dart';
+
+/// Returns `true` if the boolean [value] is `false`.
+/// Use [BoolExtensions] instead
 ///
 /// `since 0.0.1`
 bool isFalse(bool? value) => value == false;