+++ lib/src/objects.dart
@@ -6,25 +6,25 @@ import 'booleans.dart';
 /// `since 0.0.1`
 T identity<T>(T x) => x;
 
-/// Returns `true` if the dynamic [value] is `null`
+/// Returns `true` if the [value] is `null`
 ///
 /// `since 0.0.1`
-bool isNull(dynamic value) => value == null;
+bool isNull(Object? value) => value == null;
 
-/// Returns `true` if the dynamic [value] is not `null`
+/// Returns `true` if the [value] is not `null`
 ///
 /// `since 0.0.1`
-bool isNotNull(dynamic value) => isFalse(isNull(value));
+bool isNotNull(Object? value) => isFalse(isNull(value));
 
 /// If [x] is an instance of [T], returns [x] cast to [T],
 /// otherwise returns `null`.
 ///
 /// `since 0.0.1`
-T? cast<T>(dynamic x) => x is T ? x : null;
+T? cast<T>(Object? x) => x is T ? x : null;
 
 /// Calls `toString` on [x]
 ///
 /// `since 0.0.1`
-String? toString(dynamic x) {
+String? toString(Object? x) {
   return x?.toString();
 }